Refac: to have login inside the login component rather than signup, and that it would rerender the main page rather than the login component
